Always recompute the profile directory
Persisting the profile directory led to problems with the snap package
because the directory the snap is allowed to write to changes with
each new snap version.
Since there is currently no use in persisting this path anyway, we
ignore it. The respective field in the `Profile` struct is not removed
either to guarantee backwards-compatibility by making sure
configuration files containing this field can still be parsed.